+++ Blockchain/Blockche/Blockche.Blockchain.Models/AccountBalance.cs
@@ -6,16 +6,16 @@ namespace Blockche.Blockchain.Models
 {
     public class AccountBalance
     {
-        public AccountBalance(int safeBalance = 0, int confirmedBalance = 0, int pendingBalance = 0)
+        public AccountBalance(long safeBalance = 0, long confirmedBalance = 0, long pendingBalance = 0)
         {
             SafeBalance = safeBalance;
             ConfirmedBalance = confirmedBalance;
             PendingBalance = pendingBalance;
         }
 
-        public int SafeBalance { get; set; }
-        public int ConfirmedBalance { get; set; }
-        public int PendingBalance { get; set; }
+        public long SafeBalance { get; set; }
+        public long ConfirmedBalance { get; set; }
+        public long PendingBalance { get; set; }
 
     }
 }
